Classwork 6
Started by opening a text file containing a paragraph and breaking it down into individual sentences using periods as markers. Then, each sentence undergoes tokenization, where non-alphabetic characters are removed and the sentence is split into words.
To measure the similarity between sentences, a function called sentence_similarity is defined. This function converts the sentences to lowercase, creates a list of unique words, and constructs vector representations for each sentence based on word counts. The similarity between the vectors of two sentences is then calculated and subtracted from 1 to yield a similarity score. Using these similarity scores, a similarity matrix is created by comparing all pairs of sentences. This matrix is transformed into a NetworkX graph object, and the PageRank algorithm is employed to rank the sentences according to their similarity scores.
Then user is asked to specify the desired number of sentences for the summary. The top-ranked sentences, based on the user input, are selected and combined to form the final summary.
Same process is repeated for multiple text files (Text 2, Text 3, Text 4, Text 5), resulting in summaries generated for each file.