Homework 5
The process begins by loading a CSV file containing text data labeled with various emotions like joy, fear, anger, etc. Using pandas, the file is read, and a glimpse of 10 rows is displayed to give an overview of the dataset.
After that, important libraries like NLTK are imported, and NLTK's opinion lexicon and punkt tokenizer are downloaded for further use.
A word dictionary named word_dict is established where words from the NLTK opinion lexicon act as keys, and their corresponding sentiment scores (+1 for positive words and -1 for negative words) serve as the values.
After that, a function called bing_liu_score is defined. It takes input text, breaks it into words, and computes a sentiment score by summing up the sentiment values from word_dict. This scoring function is then applied to the 'Text' column of the dataset, resulting in a new column 'Bing_Liu_Score' that holds the calculated sentiment scores for each text entry.
Finally, the data is grouped by 'Emotion', and the average 'Bing_Liu_Score' for each emotion category is computed. This analysis offers insights into the average sentiment associated with each emotion represented in the dataset.